+++ packages/koot-webpack/factory-config/_defaults/isomorphic.server.prod.js
@@ -1,15 +1,17 @@
-const webpack = require('webpack')
-const common = require('../common')
+const webpack = require('webpack');
+const common = require('../common');
 
-const factoryConfig = async ({
-    pathRun,
-    // CLIENT_DEV_PORT,
-}) => {
-
+const factoryConfig = async (
+    {
+        pathRun
+        // CLIENT_DEV_PORT,
+    },
+    kootConfig = {}
+) => {
     // let { RUN_PATH, CLIENT_DEV_PORT, APP_KEY } = opt
 
     return {
-        mode: "development",
+        mode: 'development',
         devtool: 'source-map',
         target: 'async-node',
         node: {
@@ -19,17 +21,17 @@ const factoryConfig = async ({
         output: {
             filename: '[name].js',
             chunkFilename: 'chunk.[chunkhash].js',
-            path: `${pathRun}/${common.outputPath}/server`,
+            path: `${pathRun}/${common.outputPath}/server`
             // publicPath: `/[need_set_in_app:__webpack_public_path__]/`,
             // publicPath: `/`,
         },
         plugins: [
             new webpack.DefinePlugin({
-                __SPA__: false,
-            }),
+                __SPA__: false
+            })
         ],
-        externals: common.filterExternalsModules(),
-    }
-}
+        externals: common.filterExternalsModules(kootConfig)
+    };
+};
 
-module.exports = async (opt) => await factoryConfig(opt)
+module.exports = async (...args) => await factoryConfig(...args);