+++ packages/koot-webpack/libs/is-server-bundling-all-modules.js
@@ -1,8 +1,4 @@
 /**
  * 检查当前环境服务器打包结果是否需要打入 module
  */
-module.exports = () =>
-    Boolean(
-        process.env.WEBPACK_BUILD_ENV !== 'dev' &&
-            process.env.WEBPACK_BUILD_TYPE !== 'spa'
-    );
+module.exports = ({ serverPackAll } = {}) => serverPackAll === true;