+++ packages/koot/libs/validate-config/transform-compatible/webpack-related.js
@@ -10,6 +10,7 @@
  * - devDll
  * - devHmr
  * - distClientAssetsDirName
+ * - serverPackAll
  * @async
  * @param {Object} config
  * @void